Calculate the PMM log-likelihood for a given tree, data and model parameters using the Rcpp module

Calculate the PMM log-likelihood for a given tree, data and model parameters using the Rcpp module

PMMLogLikCpp(x, tree, x0, sigma2, sigmae2, cppObject = NewPMMCppObject(x,
  tree), mode = getOption("SPLITT.postorder.mode", 0))

Arguments

x
a numerical vector of size N, where N is the number of tips in tree

tree
a phylo object

x0, sigma2, sigmae2
parameters of the PMM:

x0

value at the root of tree excluding white noise;

sigma2

unit-time variance increment of the heritable component;

sigmae2

variance of the non-heritable component.

cppObject
a previously created object returned by NewPMMCppObject

mode
an integer denoting the mode for traversing the tree, i.e. serial vs parallel.

Value

the log-likelihood value.